+++ bevy_kot_ui/src/interaction/pipeline.rs
@@ -28,9 +28,9 @@ struct InteractionSourceInfoPack
 //-------------------------------------------------------------------------------------------------------------------
 
 //todo: how to handle UiTrees occluding each other?
-//todo: custom system param filter to get only widgets associated with ui in focused window ?? UIInFocusedWindow<LunexUI>
+//todo: custom system param filter to get only widgets associated with ui in focused window ?? UIInFocusedWindow<LunexUi>
 fn try_get_interaction_source_info_pack<S: InteractionSource>(
-    ui              : Query<&UiTree, With<S::LunexUI>>,  //todo: InFocusedWindow
+    ui              : Query<&UiTree, With<S::LunexUi>>,  //todo: InFocusedWindow
     cursor_pos      : CursorPos<S::LunexCursor>,
     source          : Res<S>,
     source_param    : StaticSystemParam<S::SourceParam>,
@@ -39,15 +39,15 @@ fn try_get_interaction_source_info_pack<S: InteractionSource>(
     home_zone_param : StaticSystemParam<<<S as InteractionSource>::LunexCursor as LunexCursor>::HomeZoneParam>,
     barrier_widgets : Query<
         (Entity, &Widget),
-        Or<(With<UIInteractionBarrier<S::LunexUI>>, With<InteractionBarrier<S::LunexUI, S::LunexCursor>>)>
+        Or<(With<UIInteractionBarrier<S::LunexUi>>, With<InteractionBarrier<S::LunexUi, S::LunexCursor>>)>
     >,
     unpressed_elements: Query<
         (Entity, &Widget),
-        (Without<Pressed>, With<ElementInteractionTargeter<S::LunexUI, S::LunexCursor>>)
+        (Without<Pressed>, With<ElementInteractionTargeter<S::LunexUi, S::LunexCursor>>)
     >,
     pressed_elements: Query<
         (Entity, &Widget, &PressHomeZone),
-        (With<Pressed>, With<ElementInteractionTargeter<S::LunexUI, S::LunexCursor>>)
+        (With<Pressed>, With<ElementInteractionTargeter<S::LunexUi, S::LunexCursor>>)
     >,
 ) -> Option<InteractionSourceInfoPack>
 {
@@ -194,7 +194,7 @@ fn handle_just_clicked<S: InteractionSource>(
     mut commands  : Commands,
     widgets       : Query<
         &Callback<OnClick>,
-        (With<ElementInteractionTargeter<S::LunexUI, S::LunexCursor>>, With<ElementInteractionSource<S>>)
+        (With<ElementInteractionTargeter<S::LunexUi, S::LunexCursor>>, With<ElementInteractionSource<S>>)
     >,
 ){
     // check the source was just clicked
@@ -222,7 +222,7 @@ fn handle_is_clicked<S: InteractionSource>(
     mut commands  : Commands,
     widgets       : Query<
         &Callback<OnClickHold>,
-        (With<ElementInteractionTargeter<S::LunexUI, S::LunexCursor>>, With<ElementInteractionSource<S>>)
+        (With<ElementInteractionTargeter<S::LunexUi, S::LunexCursor>>, With<ElementInteractionSource<S>>)
     >,
 ){
     // check the source is clicked
@@ -250,7 +250,7 @@ fn handle_is_clicked_home<S: InteractionSource>(
     mut commands  : Commands,
     widgets       : Query<
         &Callback<OnClickHoldHome>,
-        (With<Pressed>, With<ElementInteractionTargeter<S::LunexUI, S::LunexCursor>>, With<ElementInteractionSource<S>>)
+        (With<Pressed>, With<ElementInteractionTargeter<S::LunexUi, S::LunexCursor>>, With<ElementInteractionSource<S>>)
     >,
 ){
     // check the source is clicked
@@ -275,10 +275,10 @@ fn handle_is_clicked_home<S: InteractionSource>(
 fn handle_is_clicked_away<S: InteractionSource>(
     In(info_pack) : In<InteractionSourceInfoPack>,
     mut commands  : Commands,
-    ui            : Query<&UiTree, With<S::LunexUI>>,  //todo: InFocusedWindow
+    ui            : Query<&UiTree, With<S::LunexUi>>,  //todo: InFocusedWindow
     widgets       : Query<
         (Entity, &Widget, &CallbackWith<OnClickHoldAway, bool>),
-        (With<Pressed>, With<ElementInteractionTargeter<S::LunexUI, S::LunexCursor>>, With<ElementInteractionSource<S>>)
+        (With<Pressed>, With<ElementInteractionTargeter<S::LunexUi, S::LunexCursor>>, With<ElementInteractionSource<S>>)
     >,
 ){
     // check the source is clicked
@@ -320,7 +320,7 @@ fn handle_just_unclicked<S: InteractionSource>(
     mut commands    : Commands,
     unclick_widgets : Query<
         (Entity, &CallbackWith<OnUnClick, bool>),
-        (With<Pressed>, With<ElementInteractionTargeter<S::LunexUI, S::LunexCursor>>, With<ElementInteractionSource<S>>)
+        (With<Pressed>, With<ElementInteractionTargeter<S::LunexUi, S::LunexCursor>>, With<ElementInteractionSource<S>>)
     >,
 ){
     // check the source was just unclicked
@@ -348,7 +348,7 @@ fn handle_hover<S: InteractionSource>(
     mut commands  : Commands,
     widgets       : Query<
         &Callback<OnHover>,
-        (With<ElementInteractionTargeter<S::LunexUI, S::LunexCursor>>, With<ElementInteractionSource<S>>)
+        (With<ElementInteractionTargeter<S::LunexUi, S::LunexCursor>>, With<ElementInteractionSource<S>>)
     >,
 ){
     // see if the targeted element is hovered by the cursor
@@ -375,7 +375,7 @@ fn handle_unhover<S: InteractionSource>(
     mut commands  : Commands,
     widgets       : Query<
         (Entity, &Callback<OnUnHover>),
-        (With<Hovered>, With<ElementInteractionTargeter<S::LunexUI, S::LunexCursor>>, With<ElementInteractionSource<S>>)
+        (With<Hovered>, With<ElementInteractionTargeter<S::LunexUi, S::LunexCursor>>, With<ElementInteractionSource<S>>)
     >,
 ){
     // check if there are any widgets with callbacks